use original libnoise package structure
