+++ src/main/java/net/royawesome/jlibnoise/module/generator/Cylinders.java
@@ -23,24 +23,26 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package net.royawesome.jlibnoise.module.source;
+package net.royawesome.jlibnoise.module.generator;
+
+import com.flowpowered.math.GenericMath;
 
 import net.royawesome.jlibnoise.module.Module;
 
-public class Const extends Module {
-    public static final double DEFAULT_VALUE = 0;
-    private double value = DEFAULT_VALUE;
+public class Cylinders extends Module {
+    public static final double DEFAULT_CYLINDERS_FREQUENCY = 1.0;
+    private double frequency = DEFAULT_CYLINDERS_FREQUENCY;
 
-    public Const() {
+    public Cylinders() {
         super(0);
     }
 
-    public double getValue() {
-        return value;
+    public double getFrequency() {
+        return frequency;
     }
 
-    public void setValue(double value) {
-        this.value = value;
+    public void setFrequency(double frequency) {
+        this.frequency = frequency;
     }
 
     @Override
@@ -50,6 +52,15 @@ public class Const extends Module {
 
     @Override
     public double getValue(double x, double y, double z) {
-        return value;
+        double z1 = z;
+        double x1 = x;
+        x1 *= frequency;
+        z1 *= frequency;
+
+        double distFromCenter = Math.sqrt(x1 * x1 + z1 * z1);
+        double distFromSmallerSphere = distFromCenter - GenericMath.floor(distFromCenter);
+        double distFromLargerSphere = 1.0 - distFromSmallerSphere;
+        double nearestDist = Math.min(distFromSmallerSphere, distFromLargerSphere);
+        return 1.0 - (nearestDist * 4.0); // Puts it in the -1.0 to +1.0 range.
     }
 }
