+++ src/main/java/net/royawesome/jlibnoise/module/generator/Spheres.java
@@ -23,7 +23,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package net.royawesome.jlibnoise.module.source;
+package net.royawesome.jlibnoise.module.generator;
 
 import com.flowpowered.math.GenericMath;
